backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.12.5
pandas==2.3.3
langchain==1.1.3
langchain-google-genai==4.0.0
langchain-core==1.1.3
langchain-ollama==1.0.0
tqdm==4.67.1
python-multipart==0.0.20
python-dotenv==1.2.1
aiohttp==3.13.2
gspread==6.1.4
oauth2client==4.1.3
